Separate omc primitive types as primitive_types
